requirements.txt
torch
torchvision
pillow
exifread
numpy
